# Making int.Parse("") faster