@mikedn 2017-06-09
Have you tried using separate loops instead of a single loop and "state"?