@alexandrnikitin 2017-06-09
@mikedn Yeah, there's definitely room for optimizations!! I just wanted to initiate a discussion. If it's worth to do it (it's not only `int` related), I'll make it as fast as I can :)