@varocarbas 2017-06-09
Note that I did some quick experiments on these lines some months ago and my conclusion was that this alternative wasn't worth pursuing (didn't try it too hard though). I think that I mostly focused on `long`, but these conclusions should hold here too: very slight improvements only by analysing the `long` cases (+ irregular behaviours better/worse depending upon the specific sub-scenario); and even what seemed slightly slower performances in other cases, provoked by the inclusion of more functions in mscorlib (note that I intended to somehow emulate the existing structure, formed by various methods calling one to the other).

I am also in a position to say that further improving the main existing algorithm would be quite difficult, as far as it is already over-optimised (even though, some parts might look like easily optimisable at first sight).

I am not trying to demotivate you, but to give some context about the kind of difficulties which you are about to face. As a first step (I learned that the hard way, mostly thanks to @jkotas contributions :)), you should set up a reliable enough testing framework (in principle, many iterations and quite simplistic actions) to get a proper feeling about what you really have (is your table referring to fractions of nanosecond improvements?!).