@Pzixel 2017-06-09
It could be much better if `System.Numerics.Vector` was referenced by corlib by default. Such things are very profitable from SIMD operations (we could make up to 8 digits per clock), however afaik we are unable to use `Vector` class in such basic classes.