@jamesqo 2017-06-19
@Pzixel There is an issue for being able to use `Vector` in corelib: https://github.com/dotnet/coreclr/issues/8104. Unfortunately, it would require significant infrastructure changes so there needs to be perf measurements to determine it's worth it.