@Pzixel 2017-06-19
@jamesqo AFAIR Sasha Goldshtein @goldshtn showed that SIMD instricts can speedup operations such as string comparasion and GetHashCode in 4-8 times. It's quite good and doesn't require any specific hacks.