@varocarbas 2017-06-19
@Pzixel You should better took a look at the code you are talking about before suggesting scenarios which are very far away from being possible.

Although with Roslyn, a big proportion of the .NET code has been converted into C#/VB.NET, there are still big parts written in C/C++. The code being discussed here is one of these parts. The reason why these parts are written in these others languages is that they are much more efficient for performing certain actions (e.g., string managing).

The option you are proposing (using complex features of the given language to design the basic parts of that language), by assuming that it is even possible (i.e., those functions not depending directly upon the parts of the code which are being improved), is impractically complex/slow/will never be considered. Even in case of using the most basic functionalities of any .NET language (e.g., just loops, `Substring` and similar) the performance would be much slower. I invite you to test it yourself: download the referred code, replace all the unsafe parts (= pointers & C) with managed alternatives and see the differences. 

As said, the code being discussed is so highly optimised that improving it in a relevant way (certainly without even thinking about moving outside unsafe code) is very unlikely. Logically, you don't have to blindly believe my words, that's why I invite you to do some tests by your own, as I did, what IMHO is a basic requirement before suggesting anything anywhere (= adequately understanding what you are talking about).