@varocarbas 2017-06-19
@karelz As I think that you should already know, I am not interested in participating in certain discussions regarding the rules/expectations here (let's just say that I am knowledgeable, honest and fair and this doesn't seem my place). I also try to not participate in any discussion here, note that I was expressly referred at the start of this thread. I haven't said anything to clearly wrong and misleading statements for a while, until I finally wrote a clear correction.

On the other hand, I cannot tolerate what, sorry about that, but is a lie. I didn't insult anyone. I also shared quite relevant knowledge which apparently some people here don't seem to posses. Can you please quote the exact part of my comment where I insulted anyone? Is correcting a clearly wrong statement insulting now? And what about the remaining parts of my comment? Is it better to let people posting ideas which are clearly wrong?

I am waiting for your clarifications and/or apology.