@varocarbas 2017-06-19
@Pzixel Have you felt insulted by my last comment? Because this is what @karelz seems to be insinuating. 

If this is the case, you should re-read my comment (well... next time before someone arbitrarily deletes it) and ask me for as much help as you need to understand each single bit, because insulting you was certainly not my intention. I will not apologise/tolerate unfair reactions of people arbitrarily feeling offended by anything I say or do. Much less when this is supposed to be a technically-focused discussion and my comment was exclusively focused on technical aspects (+ your evident lack of understanding of what is being discussed here).