@karelz 2017-06-19
@varocarbas your comment was deleted as a violation of the [.NET Foundation Code of Conduct](https://dotnetfoundation.org/code-of-conduct) ([adopted by CoreCLR repo](https://github.com/dotnet/coreclr#contributing-to-repository)) as it was insulting/derogatory. You may consider this an official warning.