@karelz 2017-06-19
@varocarbas we'll be happy to share with you your original comments offline and explain in more details which parts of your comments are insulting/derogatory. Please reach out to us via email at .NET Foundation Code of Conduct (conduct@dotnetfoundation.org).

In case of Code of Conduct violations, we delete entire post. No exceptions. We want to create safe and creative environment for our community, where people are respectful to each other.

@varocarbas please do not continue discussing this issue here. Please reach out to us offline, so that we can discuss acceptable behavior according to our [Code of Conduct](https://dotnetfoundation.org/code-of-conduct).