@varocarbas 2017-06-19
@karelz I understand (in you are interested, here you have a summary of my understanding: https://plus.google.com/112097259447743562053/posts/9kt39tmAZ9K). I will not send you any email, because I know perfectly what is happening here (= this is not my place) and the chances I have to change your behaviour (= none). Feel free to delete this post and my previous one to you. I mean... you can also delete my first comment in case that you think that anyone might feel offended by it. Bye.